+++ first-person-camera.js
@@ -1,6 +1,7 @@
-// Defines a first person camera with 
+// Defines a first person camera that can be controlled by the mouse
 import { tiny } from "./examples/common.js";
 const {vec3, Mat4} = tiny
+
 export default class FirstPersonCamera {
   constructor(x, y, z) {
     this.x = x;
@@ -11,28 +12,13 @@ export default class FirstPersonCamera {
     
     this.theta = 0;
     this.phi = 0;
-    this.mouse_prev_x = 0;
-    this.mouse_prev_y = 0;
   }
 
-  rotateZ(theta) {
-    return Mat4.rotation(theta, 0, 0, 1);
-  }
-  rotateY(phi) {
-    return Mat4.rotation(phi, 0, 1, 0);
-  }
-
-
-  // Updates the camera's position based on the mouse's position
-  update_view(mouse_x, mouse_y, sensitivity) {
-    this.theta += (mouse_x - this.mouse_prev_x) * (sensitivity + 2);
-    this.phi += (mouse_y - this.mouse_prev_y) * (sensitivity + 2);
-    this.mouse_prev_x = mouse_x;
-    this.mouse_prev_y = mouse_y;
-
-    this.at = vec3(this.theta, this.phi, 1);
+  update_view(del_x, del_y, sensitivity) {
+    this.theta += del_x * (sensitivity * 0.75) / 50000;
+    this.phi += (-1) * del_y * (sensitivity * 0.75) / 50000;
+    this.at = vec3(this.theta, this.phi, 1)
     this.lookAt = Mat4.look_at(vec3(this.x, this.y, this.z), this.at, vec3(0, 1, 0));
-
     return this.lookAt;
   }
 }
